Refactor stock only TA functions
