+++ gamestonk_terminal/stocks/technical_analysis/finviz_view.py
@@ -1,67 +1,39 @@
 """ Finviz View """
 __docformat__ = "numpy"
 
-import argparse
+
 import io
-from typing import List
-import requests
 import matplotlib.pyplot as plt
-from finvizfinance.quote import finvizfinance
-from finvizfinance.util import headers
+
 from PIL import Image
 
 from gamestonk_terminal.helper_funcs import (
-    parse_known_args_and_warn,
     plot_autoscale,
 )
+from gamestonk_terminal.stocks.technical_analysis import finviz_model
 from gamestonk_terminal.config_plot import PLOT_DPI
 from gamestonk_terminal import feature_flags as gtff
 
 
-def view(other_args: List[str], ticker: str):
-    """View historical price with trendlines. [Source: Finviz]
+def view(ticker: str):
+    """View finviz image for ticker
 
     Parameters
     ----------
-    other_args : List[str]
-        argparse other args
-    ticker: str
-        stock ticker
+    ticker : str
+        Stock ticker
     """
-    parser = argparse.ArgumentParser(
-        add_help=False,
-        formatter_class=argparse.ArgumentDefaultsHelpFormatter,
-        prog="view",
-        description="""
-            View historical price with trendlines. [Source: Finviz]
-        """,
-    )
+    image_data = finviz_model.get_finviz_image(ticker)
+    dataBytesIO = io.BytesIO(image_data)
+    im = Image.open(dataBytesIO)
 
-    try:
-        ns_parser = parse_known_args_and_warn(parser, other_args)
-        if not ns_parser:
-            return
+    fig, ax = plt.subplots(figsize=plot_autoscale(), dpi=PLOT_DPI)
+    ax.set_axis_off()
+    fig.subplots_adjust(left=0, right=1, top=1, bottom=0)
 
-        stock = finvizfinance(ticker)
-        image_url = stock.TickerCharts(urlonly=True)
+    if gtff.USE_ION:
+        plt.ion()
 
-        r = requests.get(image_url, stream=True, headers=headers, timeout=5)
-        r.raise_for_status()
-        r.raw.decode_content = True
-
-        dataBytesIO = io.BytesIO(r.content)
-        im = Image.open(dataBytesIO)
-
-        fig, ax = plt.subplots(figsize=plot_autoscale(), dpi=PLOT_DPI)
-        ax.set_axis_off()
-        fig.subplots_adjust(left=0, right=1, top=1, bottom=0)
-
-        if gtff.USE_ION:
-            plt.ion()
-
-        plt.imshow(im)
-        plt.show()
-        print("")
-
-    except SystemExit:
-        print("")
+    plt.imshow(im)
+    plt.show()
+    print("")
